requirements.txt
easydict==1.9
einops==0.8.1
h5py==3.9.0
matplotlib==3.7.5
munch==4.0.0
numpy==1.24.4
open3d==0.13.0
opencv_python==4.8.0.76
pandas==1.3.4
Pillow==11.3.0
plyfile==0.7.4
pointnet2_ops==3.0.0
pyvista==0.42.1
PyYAML==5.4.1
pyzed==4.0
pyzedhub==0.73
scikit_learn==0.24.1
scipy==1.7.1
setuptools==52.0.0.post20210125
tensorboard_logger==0.1.0
termcolor==3.1.0
timm==0.4.5
torch==2.0.0
tqdm==4.59.0
transforms3d==0.4.1
ultralytics==8.3.161

